+++ examples/builder/internal-composable.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
-require 'callable_tree'
+require_relative '../../lib/callable_tree'
+# require 'callable_tree'
 
 def less_than(num)
   # The following block call is equivalent to calling `super` in the class style.